requirements.txt
tensorflow
pandas
matplotlib
pillow
seaborn
scikit-learn
pathlib
numpy
keras-tuner
datasets